backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
gunicorn==21.2.0
groq==0.4.2
google-generativeai==0.3.0
newspaper3k==0.2.8
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
pymongo==4.5.0
aiohttp==3.9.1
newsapi-python==0.2.7
vt-py==0.15.0